Include file cleanup and fix timeout bug

+++ scripts/anki-media.py
@@ -14,28 +14,39 @@ MACRO_OUTPUT_FILE = '/Users/ptran/Music/macro-output/anki.mp3'
 client = pipeclient.PipeClient()
 
 def prepare_macro_input(filename):
-        shutil.copyfile(SOURCE_DIR + '/' + filename, MACRO_INPUT_FILE)
+	shutil.copyfile(SOURCE_DIR + '/' + filename, MACRO_INPUT_FILE)
 
-def execute_macro(macro, timeout=1):
-        start = time.time()
-        client.write('Macro_' + macro.replace(' ', ''))
+def execute_macro(macro, timeout=5):
+	start = time.time()
+	client.write('Macro_' + macro.replace(' ', ''))
 
-        reply = ''
-        while reply == '' or time.time() - start > timeout:
-                time.sleep(0.1)
-                reply = client.read()
+	reply = ''
+	while reply == '' and time.time() - start < timeout:
+		# print("looping duration: ", time.time() - start)
+		time.sleep(0.5)
+		reply = client.read()
+		# print("Reply: ", reply)
+	if time.time() - start >= timeout:
+		print("Skipping..")
+
 
 def copy_macro_output(filename):
-        shutil.copyfile(MACRO_OUTPUT_FILE, TARGET_DIR + '/' + filename)
+	if os.path.exists(MACRO_OUTPUT_FILE):
+		shutil.copyfile(MACRO_OUTPUT_FILE, TARGET_DIR + '/' + filename)
+
+def clean_output_file():
+	if os.path.exists(MACRO_OUTPUT_FILE):
+		os.remove(MACRO_OUTPUT_FILE)
 
 def main():
-        for filename in os.listdir(SOURCE_DIR):
-                if not filename.endswith('mp3'):
-                        continue
-                print("Processing: ", filename)
-                prepare_macro_input(filename)
-                execute_macro('Anki Script')
-                copy_macro_output(filename)
-        print("Script completed")
+	for filename in os.listdir(SOURCE_DIR):
+		if not filename.endswith('mp3'):
+			continue
+		print("Processing: ", filename)
+		prepare_macro_input(filename)
+		execute_macro('Anki Script')
+		copy_macro_output(filename)
+		clean_output_file()
+	print("Script completed")
 
 main()
